requirements.txt
Flask
Flask-Bcrypt
Flask-SQLAlchemy
Flask-Login
Flask-CORS
pyfatsecret
requests
python-dotenv
Flask-Migrate==4.0.5
alembic==1.13.1
